fix the bug of docker filesystem and fix test cases 1. fix the bug that docker filesystem can't append file and read file 2. fix the test cases because different test cases share the same path by using tmp_path for each test cases 3. make the scope of container fixture into class

+++ jupyterMagicCommands/utils/docker.py
@@ -1,8 +1,11 @@
-import tarfile
 import os
+import tarfile
+import tempfile
 from docker.models.containers import Container
 
 def copy_to_container(container: Container, src: str, dst: str) -> None:
+    if not os.path.exists(src):
+        raise FileNotFoundError(f"Source file {src} can't be found")
     with tarfile.open(src + '.tar', mode='w') as f:
         f.add(src, arcname=os.path.basename(dst))
     with  open(src + '.tar', 'rb') as f:
@@ -10,3 +13,22 @@ def copy_to_container(container: Container, src: str, dst: str) -> None:
     targetFolder = os.path.dirname(dst)
     container.exec_run(f"mkdir -p '{targetFolder}'")
     container.put_archive(targetFolder, data)
+
+def _rename_members(members, original, target):
+    for member in members:
+        if member.name.startswith(original):
+            member.name = target + member.name[len(original):]
+
+def copy_from_container(container: Container, src: str, dst: str) -> None:
+    # make sure target folder exists
+    targetFolder = os.path.dirname(dst)
+    if not os.path.exists(targetFolder):
+        os.makedirs(targetFolder)
+    bits, stat = container.get_archive(src)
+    with tempfile.NamedTemporaryFile('wb+', delete=False) as f:
+        for chunk in bits:
+            f.write(chunk)
+    with tarfile.open(f.name) as tar:
+        _rename_members(tar.getmembers(), os.path.basename(src), os.path.basename(dst))
+        # path参数指定解压到的目录
+        tar.extractall(path=targetFolder)
